+++ learn-designpatterns-in-csharp/Source/Behavioral/Memento/EmployeeManagerDemo/Receivers/IEmployeeManagerRepository.cs
@@ -1,0 +1,16 @@
+﻿using EmployeeManagerDemo.Models;
+
+namespace EmployeeManagerDemo.Receivers
+{
+    /// <summary>
+    /// Receiver (interface)
+    /// </summary>
+    public interface IEmployeeManagerRepository
+    {
+        void AddEmployee(int managerId, Employee employee);
+        void RemoveEmployee(int managerId, Employee employee);
+        bool HasEmployee(int managerId, int employeeId);
+        void WriteDataStore();
+    }
+
+}
